django_defaults/requirements.txt
Django==1.8.4
dj-database-url==0.3.0
django-crispy-forms==1.5.2
